+++ src/Business/Grand.Business.Checkout/Commands/Models/Orders/PartiallyRefundOfflineCommand.cs
@@ -6,6 +6,6 @@ namespace Grand.Business.Checkout.Commands.Models.Orders
     public class PartiallyRefundOfflineCommand : IRequest<bool>
     {
         public PaymentTransaction PaymentTransaction { get; set; }
-        public decimal AmountToRefund { get; set; }
+        public double AmountToRefund { get; set; }
     }
 }
