+++ src/Business/Grand.Business.Checkout/Queries/Models/Orders/CanPartiallyPaidOfflineQuery.cs
@@ -6,6 +6,6 @@ namespace Grand.Business.Checkout.Queries.Models.Orders
     public class CanPartiallyPaidOfflineQuery : IRequest<bool>
     {
         public PaymentTransaction PaymentTransaction { get; set; }
-        public decimal AmountToPaid { get; set; }
+        public double AmountToPaid { get; set; }
     }
 }
