+++ src/Business/Grand.Business.Checkout/Queries/Models/Orders/CanPartiallyRefundQuery.cs
@@ -5,7 +5,7 @@ namespace Grand.Business.Checkout.Queries.Models.Orders
 {
     public class CanPartiallyRefundQuery : IRequest<bool>
     {
-        public decimal AmountToRefund { get; set; }
+        public double AmountToRefund { get; set; }
         public PaymentTransaction PaymentTransaction { get; set; }
     }
 }
